Extract apply command from obliterate cog into a new 'shared' cog for re-use
